tests/wpt/tests/css/css-text/parsing/word-break-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-break with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-word-break">
<meta name="assert" content="word-break supports only the grammar 'normal | keep-all | break-all | break-word'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("word-break", "auto");
test_invalid_value("word-break", "normal keep-all");
test_invalid_value("word-break", "break-all break-all");
test_invalid_value("word-break", "normal break-word");
test_invalid_value("word-break", "break-word normal");
test_invalid_value("word-break", "keep-all break-word");
test_invalid_value("word-break", "break-all break-word");
test_invalid_value("word-break", "word-break");
</script>
</body>
</html>
