third_party/blink/web_tests/external/wpt/css/css-text/parsing/word-break-invalid-expected.txt

This is a testharness.js-based test.
FAIL e.style['word-break'] = "auto" should not set the property value assert_equals: expected "" but got "auto"
PASS e.style['word-break'] = "normal keep-all" should not set the property value
PASS e.style['word-break'] = "break-all break-all" should not set the property value
PASS e.style['word-break'] = "normal break-word" should not set the property value
PASS e.style['word-break'] = "break-word normal" should not set the property value
PASS e.style['word-break'] = "keep-all break-word" should not set the property value
PASS e.style['word-break'] = "break-all break-word" should not set the property value
PASS e.style['word-break'] = "word-break" should not set the property value
Harness: the test ran to completion.
